# swagger.yml
swagger: "2.0"
...
info:
  description: 'Group 25 Large Project - Coffee - API Documentation'
  version: '1.0.0'
  title: 'Coffee'
  contact:
    email: 'mi993440@ucf.edu'
...
securityDefinitions:
  BearerAuth:
    type: apiKey
    name: Authorization
    in: header
    description: 'JWT auth header. Format: "Bearer <token>"'

# Docs
paths:
...
  /api/auth/me:
    get:
      tags:
        - auth
      summary: Get the current user
      description: >
        Restores a session from a stored token (e.g. after a page refresh) without
        re-prompting for a password. Returns the same user/settings shape as login.
      operationId: getCurrentUser
      security:
        - BearerAuth: []
      produces:
        - application/json
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/userResponse'
        403:
          description: Forbidden -- account has been deactivated (code ACCOUNT_DEACTIVATED), or was deactivated and has since been permanently deleted for inactivity (code ACCOUNT_DELETED, no deactivatesAt field)
          schema:
            $ref: '#/definitions/deactivatedResponse'
        404:
          description: User Not Found
          schema:
            $ref: '#/definitions/errorResponse'
        500:
          description: Server Error
          schema:
            $ref: '#/definitions/errorResponse'
  /api/auth/verify-email/{token}:
    get:
      tags:
        - auth
      summary: Verify an account via its emailed link
      description: >
        Meant to be opened directly from the verification email, not called as a JSON API --
        it renders a plain HTML confirmation page rather than returning JSON. The token is a
        one-time value that expires 24 hours after registration.
      operationId: verifyEmail
      produces:
        - text/html
      parameters:
        - in: path
          name: token
          required: true
          type: string
      responses:
        200:
          description: OK -- account verified. Returns an HTML confirmation page.
          schema:
            type: string
        400:
          description: Bad Request -- token invalid or expired. Returns an HTML error page.
          schema:
            type: string
        500:
          description: Server Error. Returns an HTML error page.
          schema:
            type: string
  /api/auth/resend-verification:
    post:
      tags:
        - auth
      summary: Resend the account verification email
      description: >
        Issues a new verification token (invalidating the previous one) and re-sends it.
        Rate-limited to once per 60 seconds per account, enforced server-side regardless
        of any client-side cooldown.
      operationId: resendVerification
      security:
        - BearerAuth: []
      produces:
        - application/json
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/messageResponse'
        400:
          description: Bad Request -- account is already verified
          schema:
            $ref: '#/definitions/errorResponse'
        404:
          description: User Not Found
          schema:
            $ref: '#/definitions/errorResponse'
        429:
          description: Too Many Requests -- cooldown still active
          schema:
            $ref: '#/definitions/resendCooldownResponse'
        500:
          description: Server Error
          schema:
            $ref: '#/definitions/errorResponse'
  /api/auth/banner:
    patch:
      tags:
        - auth
      summary: Update the account banner image
      description: Sets (or clears, with an empty string) the account's banner image. Pass the relative path returned by /api/items/upload.
      operationId: updateBanner
      security:
        - BearerAuth: []
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/updateBannerRequest'
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/userResponse'
        404:
          description: User Not Found
          schema:
            $ref: '#/definitions/errorResponse'
        500:
          description: Server Error
          schema:
            $ref: '#/definitions/errorResponse'
  /api/auth/settings:
    patch:
      tags:
        - auth
      summary: Update account dashboard settings
      description: >
        Partially updates the account's company/branding and low-stock notification
        settings. Only the fields present in the body are changed; anything outside
        the whitelisted fields is silently ignored.
      operationId: updateSettings
      security:
        - BearerAuth: []
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/updateSettingsRequest'
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/userResponse'
        400:
          description: Bad Request (invalid/out-of-range field, or no valid fields provided)
          schema:
            $ref: '#/definitions/errorResponse'
        404:
          description: User Not Found
          schema:
            $ref: '#/definitions/errorResponse'
        500:
          description: Server Error
          schema:
            $ref: '#/definitions/errorResponse'
  /api/auth/account:
    delete:
      tags:
        - auth
      summary: Delete the logged-in account
      description: Permanently deletes the account along with all of its items, categories, and uploaded files (item photos, banner). Cannot be undone.
      operationId: deleteAccount
      security:
        - BearerAuth: []
      produces:
        - application/json
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/messageResponse'
        404:
          description: User Not Found
          schema:
            $ref: '#/definitions/errorResponse'
        500:
          description: Server Error
          schema:
            $ref: '#/definitions/errorResponse'
...
definitions:
...
  accountSettings:
    type: object
    properties:
      companyName:
        type: string
        example: Coffee Hour
      businessType:
        type: string
        example: Coffee shop
      managerName:
        type: string
        example: Alex Morgan
      accentColor:
        type: string
        description: Hex color code
        example: '#a9642e'
      notificationsEnabled:
        type: boolean
        example: false
      notificationFrequency:
        type: string
        enum: [immediate, hourly, daily]
        example: immediate
  updateSettingsRequest:
    type: object
    description: All fields are optional -- only the ones present in the body are updated.
    properties:
      companyName:
        type: string
        example: Coffee Hour
      businessType:
        type: string
        example: Coffee shop
      managerName:
        type: string
        example: Alex Morgan
      accentColor:
        type: string
        description: Hex color code
        example: '#a9642e'
      notificationsEnabled:
        type: boolean
        example: true
      notificationFrequency:
        type: string
        enum: [immediate, hourly, daily]
        example: hourly
  updateBannerRequest:
    type: object
    required:
      - bannerImage
    properties:
      bannerImage:
        type: string
        description: Relative /uploads/... path from /api/items/upload, or an empty string to remove the banner.
        example: /uploads/1a2b3c.jpg
  user:
    type: object
    properties:
      id:
        type: string
        example: '65123abc4567def890123456'
      email:
        type: string
        example: jdoe@email.com
      isVerified:
        type: boolean
        example: false
      bannerImage:
        type: string
        example: /uploads/1a2b3c.jpg
      deactivatesAt:
        type: string
        format: date-time
        description: When this account will be auto-deactivated. Only present on login and /me responses; null if the account isn't verified yet.
        example: '2026-07-27T00:00:00.000Z'
      settings:
        $ref: '#/definitions/accountSettings'
...
  userResponse:
    type: object
    properties:
      user:
        $ref: '#/definitions/user'
      error:
        type: string
        example: ""

  # ---------------- Categories ----------------
...
  messageResponse:
    type: object
    properties:
      message:
        type: string
        example: OK
      error:
        type: string
        example: ""
  errorResponse:
    type: object
    properties:
      error:
        type: string
        example: Something went wrong.
      details:
        type: string
        example: Additional error detail from the server, when available.
  resendCooldownResponse:
    type: object
    properties:
      error:
        type: string
        example: Please wait before requesting another verification email
      retryAfterSeconds:
        type: integer
        example: 42
  deactivatedResponse:
    type: object
    properties:
      error:
        type: string
        example: This account has been deactivated.
      code:
        type: string
        example: ACCOUNT_DEACTIVATED
      deactivatesAt:
        type: string
        format: date-time
        example: '2026-07-27T00:00:00.000Z'